% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asSurvey.R
\name{as_survey}
\alias{as_survey}
\title{Creates basic survey table from votes in percent
This functions takes votes in percent (per party) obtained from a survey, 
and returns a table containing votes (in percent) and party names. Conducts  
sanity checks along the way, such as checking that percentages add up to 1.}
\usage{
as_survey(votes.in.perc, samplesize, parties = c("CDU/CSU", "SPD", "GRUENE",
  "FDP", "LINKE", "PIRATEN", "AfD", "FW", "Sonstige"), epsilon = 1e-05)
}
\arguments{
\item{votes.in.perc}{Votes in percent each party received in the survey 
of interest. Can be set to \code{NA}, if parties are specified that are not 
mentioned in the specific survey (otherwise the \code{parties} argument has 
to be modified).}

\item{samplesize}{Number of respondents in survey.}

\item{parties}{Vector of same length and in the same orderas \code{votes.in.perc}}

\item{epsilon}{The parameter \code{votes.in.perc} should add up to one. 
This parameter controls the maximal numerical divergence allowed.}
}
\value{
A \code{data.frame} containing input and absolute number of votes 
in survey per party.
}
\description{
Creates basic survey table from votes in percent
This functions takes votes in percent (per party) obtained from a survey, 
and returns a table containing votes (in percent) and party names. Conducts  
sanity checks along the way, such as checking that percentages add up to 1.
}
\seealso{
\code{\link{redistribute}}
}
\keyword{survey}

