% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-helpers.R
\name{collapse_parties}
\alias{collapse_parties}
\title{Transform surveys in long format}
\usage{
collapse_parties(
  surveys,
  parties = c("cdu", "spd", "greens", "fdp", "left", "pirates", "fw", "afd", "others")
)
}
\arguments{
\item{surveys}{A data frame with one survey per row.}
}
\value{
Data frame in long format
}
\description{
Given a data frame containing multiple surveys (one row per survey), transforms
the data into long format with one row per party.
}
\examples{
emnid <- scrape_wahlrecht()
emnid.long <- collapse_parties(emnid)
}
