% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{cprob_row}
\alias{cprob_row}
\title{Coalition probabilities based on one survey}
\usage{
cprob_row(row, nsim = 10000L, coalitions = list(c("CDU"), c("CDU", "FDP"),
  c("CDU", "FDP", "GRUENE"), c("SPD"), c("SPD", "LINKE"), c("SPD", "LINKE",
  "GRUENE")), superior.coalitions = list(c(""), c("CDU"), c("CDU", "FDP"),
  c(""), c("SPD"), c("SPD", "LINKE")), majority = 300L, ...)
}
\arguments{
\item{row}{A single row of a data frame. Should contain percentages of votes 
for individual parties (that add up to 100) and sample size of survey.}

\item{nsim}{number of simulations}

\item{coalitions}{A list of coaltions for which coalition probabilities should 
be calculated. Each list entry must be a vector of party names. Those names 
need to correspond to the names in \code{seat.tabs}.}

\item{superior.coalitions}{A list similar to \code{coalitions} and of the 
same length. Coalition probabilities for coalitions in \code{coalitions} list 
will be calculated excluding times where respective entry in 
\code{superior.coalitions} is also possible.}

\item{majority}{Number of seats required for majority in parliament.
Defaults to 300 (required number of seats in german parliament).}

\item{...}{Further arguments passed to \code{draw_from_posterior}}
}
\description{
Given one row from a data frame containing results of surveys (one  survey 
per row), returns specified coalition probabilities
}
