% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{cprob_tab}
\alias{cprob_tab}
\title{Coalition probabilities based on survey results}
\usage{
cprob_tab(survey.df, nsim = 10000L, coalitions = list("CDU", c("CDU",
  "FDP"), c("CDU", "FDP", "GRUENE"), "SPD", c("SPD", "LINKE"), c("SPD", "LINKE",
  "GRUENE")), superior.coalitions = list("", "CDU", c("CDU", "FDP"), "",
  "SPD", c("SPD", "LINKE")), mc.cores = 10L, ...)
}
\arguments{
\item{survey.df}{A data frame containing survey results (one survey per row).}

\item{nsim}{number of simulations}

\item{coalitions}{A list of coaltions for which coalition probabilities should 
be calculated. Each list entry must be a vector of party names. Those names 
need to correspond to the names in \code{seat.tabs}.}

\item{superior.coalitions}{A list similar to \code{coalitions} and of the 
same length. Coalition probabilities for coalitions in \code{coalitions} list 
will be calculated excluding times where respective entry in 
\code{superior.coalitions} is also possible.}

\item{mc.cores}{Number of cores used for calculations.}

\item{...}{Further arguments passed to \code{draw_from_posterior}}
}
\description{
Given  data frame containing results from different surveys (one  survey 
per row), returns the provided data frame plus specified coalition probabilities.
}
\examples{
library(dplyr)
tab <- scrape_wahlrecht() \%>\% slice(1:3)
cprob_tab(tab, nsim=100, mc.cores=1L)
}
