% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHondt.R
\name{dHondt}
\alias{dHondt}
\title{Seat Distribution by D'Hondt}
\usage{
dHondt(votes, parties, n_seats = 183)
}
\arguments{
\item{votes}{Number of votes per party.}

\item{parties}{Names of parties (must be same length as votes).}

\item{n_seats}{Number of seats in parliament. Defaults to 183 (seats in
Austrian parliament).}
}
\value{
A numeric vector containing the seats of all parties after
redistribution via D'Hondt
}
\description{
Calculates number of seats for the respective parties according to the
method of d'Hondt.
}
\examples{
library(coalitions)
library(dplyr)
# get the latest survey for a sample of German federal election polls
surveys <- get_latest(surveys_sample) \%>\% tidyr::unnest("survey")
# calculate the seat distribution based on D'Hondt for a parliament with 300 seats
dHondt(surveys$votes, surveys$party, n_seats = 300)
}
\seealso{
\code{\link{sls}}
}
