% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoalitionProbability_v3.R
\name{filter_superior}
\alias{filter_superior}
\title{Remove rows from table for which superior coalitions possible}
\usage{
filter_superior(majority_df, coalition, ...)
}
\arguments{
\item{majority_df}{A data frame containing logical values indicating 
if the coalitions (columns) have a majority (rows).}

\item{...}{Further arguments passed to \code{link[coalitions]{get_superior}}}
}
\description{
Remove rows from table for which superior coalitions possible
}
\examples{
test_df <- data.frame(
 cdu            = c(rep(F, 9), T),
 cdu_fdp        = c(rep(F, 8), T, T),
 cdu_fdp_gruene = c(T, T, rep(F, 6), T, T))
filter_superior(test_df, "cdu_fdp_gruene")
}
\seealso{
\code{\link[coalitions]{get_superior}}
}
