% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoalitionProbability.R
\name{get_coalition_probabilities}
\alias{get_coalition_probabilities}
\alias{get_probabilities}
\title{Calculate coalition probabilities}
\usage{
get_coalition_probabilities(seat.tabs, coalitions, superior.coalitions,
  majority = 300)

get_probabilities(seat.tabs, coalitions, superior.coalitions, majority = 300)
}
\arguments{
\item{seat.tabs}{A list of data frames containing simulated seat distributions.}

\item{coalitions}{A list of coaltions for which coalition probabilities should 
be calculated. Each list entry must be a vector of party names. Those names 
need to correspond to the names in \code{seat.tabs}.}

\item{superior.coalitions}{A list similar to \code{coalitions} and of the 
same length. Coalition probabilities for coalitions in \code{coalitions} list 
will be calculated excluding times where respective entry in 
\code{superior.coalitions} is also possible.}

\item{majority}{Number of seats required for majority in parliament.
Defaults to 300 (required number of seats in german parliament).}
}
\value{
Probabilities for majority for specified coalitions.
}
\description{
Given a list containing simulated seats obtained per party
}
\seealso{
get_seat_distribution
}

