% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoalitionProbability_v3.R
\name{get_probs}
\alias{get_probs}
\title{Wrapper for calculation of coalition probabilities from survey}
\usage{
get_probs(x, coalitions = list(c("cdu"), c("cdu", "fdp"), c("cdu", "fdp",
  "gruene"), c("spd"), c("spd", "linke"), c("spd", "linke", "gruene")),
  nsim = 1e+05, distrib.fun = sls2, seats_majority = 300L)
}
\arguments{
\item{x}{A table containing one row per survey and survey information in 
long format in a separate column named \code{survey}.}

\item{coalitions}{A list of coaltions for which coalition probabilities should 
be calculated. Each list entry must be a vector of party names. Those names 
need to correspond to the names in \code{majority_df}.}

\item{nsim}{number of simulations}

\item{distrib.fun}{Function to calculate seat distribution. Defaults to 
\code{\link{sls}} (Sainte-Lague/Scheppers).}

\item{seats_majority}{The number of seats needed to obtain majority.}
}
\description{
Wrapper for calculation of coalition probabilities from survey
}
\seealso{
\code{\link[coalitions]{calculate_prob}}
}
