% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeatDistribution.R
\name{get_seat_distribution}
\alias{get_seat_distribution}
\alias{get_seats}
\title{Calculate seat distribution from draws from posterior}
\usage{
get_seat_distribution(dirichlet.draws, survey, distrib.fun = sls,
  samplesize = NULL, ...)

get_seats(dirichlet.draws, survey, mc.cores = 1, distrib.fun = sls2,
  samplesize = NULL, .id = "sim", ...)
}
\arguments{
\item{dirichlet.draws}{Matrix containing random draws from posterior.}

\item{survey}{The actual survey results on which \code{dirichlet.draws}
were based on.}

\item{distrib.fun}{Function to calculate seat distribution. Defaults to 
\code{\link{sls}} (Sainte-Lague/Scheppers).}

\item{samplesize}{Number of individuals participating in the \code{survey}.}

\item{...}{Further arguments passed to \code{distrib.fun}.}

\item{mc.cores}{Number of cores to be used in parallel.
See \code{\link[parallel]{mclapply}}.}

\item{.id}{If not \code{NULL} a variable with this name will be created
giving either the name or the index of the data frame.}
}
\value{
\code{list} containing seat distributions for each simulation in 
\code{dirichlet.draws}
}
\description{
Calculate seat distribution from draws from posterior
}
\seealso{
\code{\link{draw_from_posterior}}, \code{\link{sls}}, 
\code{\link{dHondt}}
}
\keyword{distribution}
\keyword{seat}
