% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seat-distribution.R
\name{get_seats2}
\alias{get_seats2}
\title{Calculate seat distribution from draws from posterior}
\usage{
get_seats2(dirichlet.draws, survey, distrib.fun = sls2, samplesize = NULL,
  .id = "sim", hurdle = 0.05, others = "sonstige", mc.cores = 1, ...)
}
\arguments{
\item{dirichlet.draws}{Matrix containing random draws from posterior.}

\item{survey}{The actual survey results on which \code{dirichlet.draws}
were based on.}

\item{distrib.fun}{Function to calculate seat distribution. Defaults to 
\code{\link{sls}} (Sainte-Lague/Scheppers).}

\item{samplesize}{Number of individuals participating in the \code{survey}.}

\item{.id}{If not \code{NULL} a variable with this name will be created
giving either the name or the index of the data frame.}

\item{mc.cores}{Number of cores to be used in parallel.
See \code{\link[parallel]{mclapply}}.}

\item{...}{Further arguments passed to \code{distrib.fun}.}
}
\value{
\code{list} containing seat distributions for each simulation in 
\code{dirichlet.draws}
}
\description{
Calculate seat distribution from draws from posterior
}
\seealso{
\code{\link{draw_from_posterior}}, \code{\link{sls}}, 
\code{\link{dHondt}}
}
