% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{get_surveys_nds}
\alias{get_surveys_nds}
\title{Obtain (nested) Niedersachsen surveys object}
\usage{
get_surveys_nds()
}
\description{
Scrapes data from \url{wahlrecht.de} and performs some sanitizing.
}
