% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoalitionProbability_v3.R
\name{has_majority}
\alias{has_majority}
\title{Does a coalition have a majority}
\usage{
has_majority(seats_tab, coalition, seats_majority = 300L)
}
\arguments{
\item{seats_tab}{A table containing information on how many seats each party
obtained.}

\item{coalition}{The coaliton of interest for which superior coalitions 
will be obtained by \code{\link[coalitions]{get_superior}}.}

\item{seats_majority}{The number of seats needed to obtain majority.}
}
\description{
Does a coalition have a majority
}
