% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoalitionProbability_v3.R
\name{have_majority}
\alias{have_majority}
\title{Do coalitions have a majority}
\usage{
have_majority(seats_tab, coalitons = list(c("cdu"), c("cdu", "fdp"), c("cdu",
  "fdp", "gruene"), c("spd"), c("spd", "linke"), c("spd", "linke", "gruene")),
  majority = 300L, collapse = "_")
}
\arguments{
\item{seats_tab}{A data frame cotaining number of seats obtained by a party. 
Must have columns \code{party} and \code{seats}.}
}
\description{
Do coalitions have a majority
}
