% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{initialize}
\alias{initialize}
\title{Initial dowload and data base creation}
\usage{
initialize(path = "./")
}
\arguments{
\item{path}{A path to the folder in which data base should be initialized.}
}
\description{
Initial dowload and data base creation
}
