% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{optim_eff}
\alias{optim_eff}
\title{Calculate the effective sample size}
\usage{
optim_eff(size, share, corr)
}
\arguments{
\item{size}{A vecotr of sample sizes from different surveys (from different 
institutes) for one party.}

\item{share}{The relative share of votes for party of interest ([0-1])}

\item{corr}{Assumed correlation between surveys (of different institutes). 
Defaults to 0.5.}
}
\description{
This is the work horse function that calculates the effective sample size. 
Should usually not be called by the user directly.
}
