% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{party_colors_de}
\alias{party_colors_de}
\title{Colors for German parties}
\format{
A named character vector. Names indicate parties. Values contain
color strings for the respective parties
}
\usage{
party_colors_de
}
\description{
A vector of colors associated with German parties.
}
\keyword{datasets}
