% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{party_labels_de}
\alias{party_labels_de}
\title{Labels for German parties}
\format{
A named character vector. Names indicate parties. Values contain
  party names suitable for plot labels.
}
\usage{
party_labels_de
}
\description{
A vector of labels associated with German parties.
}
\keyword{datasets}
