% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survey.R
\name{plot_survey}
\alias{plot_survey}
\title{Plot voter shares observed in one survey}
\usage{
plot_survey(data, cols = NULL, labels = NULL, focus = NULL,
  base_size = 15, minimal = FALSE, ylim, ylab = "Voter share in \%",
  main = NULL, perc_labels = TRUE, hurdle = 5)
}
\arguments{
\item{data}{scraped dataset containing one row per party in the column \code{party}
and the observed voter share in the column \code{percent}}

\item{cols}{named vector containing party colours}

\item{labels}{named vector containing party labels}

\item{focus}{optional name of one party ("cdu","spd",etc.) that should be highlighted}

\item{base_size}{base size of all texts in plot, see \code{\link[ggplot2]{ggtheme}}}

\item{minimal}{if TRUE the plot axes and some other graphic elements are suppressed.
Defaults to FALSE.}

\item{ylim}{optional limits vector for the y axis}

\item{ylab}{optional label for the y axis}

\item{main}{optional plot title}

\item{perc_labels}{if TRUE (default) percentage labels are drawn above the bars}

\item{hurdle}{hurdle for single parties to get into the parliament, e.g. '5' for '5\%'.
The horizontal line can be suppressed using \code{NA}.}
}
\description{
Bar chart of the raw voter shares observed in one survey.
Additionally to plotting positive voter shares,
the function can be used to plot party-specific differences (e.g. between
a survey and the election result), including negative numbers.
}
\examples{
library(tidyverse)
dat <- scrape_wahlrecht() \%>\% collapse_parties() \%>\%
  slice(1) \%>\% select(survey) \%>\% unnest()

cols <- c("cdu" = "black", "spd" = "#E3000F", "greens" = "#46962b", "fdp" = "gold3",
          "left" = "deeppink3", "afd" = "skyblue2", "others" = "grey")
labels <- c("cdu" = "Union", "spd" = "SPD", "greens" = "Greens", "fdp" = "FDP",
            "left" = "Left", "afd" = "AfD", "others" = "Others")
plot_survey(dat, cols = cols, labels = labels)
plot_survey(dat, cols = cols, labels = labels, minimal = TRUE)
}
