% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{pool_surveys}
\alias{pool_surveys}
\title{Obtain pooled survey during specified period}
\usage{
pool_surveys(surveys, last_date = Sys.Date(), institutes = c("allensbach",
  "emnid", "forsa", "fgw", "gms", "infratest", "insa"), period = 14,
  corr = 0.5, weights = NULL)
}
\arguments{
\item{surveys}{A \code{tibble} containing survey results for multiple
institutes as returned by \code{\link[coalitions]{get_surveys}}.}

\item{last_date}{Only surveys in the time-window from \code{last_date} to
\code{last_date} - period will be considered for each institute. Defaults
to current date.}

\item{institutes}{Character vector of institutes that should be considered
for pooling.}

\item{period}{See \code{last_date} argument.}

\item{corr}{Assumed correlation between surveys (of different institutes).
Defaults to 0.5.}

\item{weights}{Additional weights for individual surveys.}
}
\description{
Per default, pools surveys starting from current date and going 14 days back.
For each institute within the defined time-frame, only the most recent survey
is used.
}
\examples{
library(coalitions)
pool_surveys(surveys_sample)
}
