% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{prettify_parties}
\alias{prettify_parties}
\alias{prettify_coalitions}
\title{Make prettier labels for parties}
\usage{
prettify_parties(x, current = c("cdu", "spd", "gruene", "fdp", "linke",
  "piraten", "afd", "sonstige"), new = c("Union", "SPD", "Gr端ne", "FDP",
  "Die Linke", "Piraten", "AfD", "Sonstige"))

prettify_coalitions(x, current = c("cdu", "cdu_fdp", "cdu_gruene",
  "cdu_fdp_gruene", "spd", "linke_spd", "gruene_linke_spd"), new = c("Union",
  "Union - FDP", "Union - Gr端ne", "Union - FDP - Gr端ne", "SPD",
  "SPD - Die Linke", "SPD - Die Linke - Gr端ne"))
}
\arguments{
\item{x}{a Vector of parties to prettify.}

\item{current}{A vector of current values \code{x} can take}

\item{new}{A vector of labels that \code{x} should be transformed into.}

\item{x}{A vector of coalitions to prettify.}
}
\description{
Make prettier labels for parties

Make prettier labels for coalitoins
}
\keyword{internal}
