% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.R
\name{prettify_strings}
\alias{prettify_strings}
\title{Replace/prettify matching words/terms in one vector by another}
\usage{
prettify_strings(x, current = .trans_df$german,
  new = .trans_df$german_pretty)
}
\arguments{
\item{x}{A character vector (or factor) that should be renamed.}

\item{current}{A vector of characters (possibly subset of \code{x}).
Entries in \code{x} that match entries in \code{current} will be renamed
accroding to entries in \code{new}.}

\item{new}{A vector of characters that will replace entries in \code{x} which
have matches in \code{current}.}
}
\description{
The function searches for \code{x} values, that occur in \code{current}
and replaces them with entries in \code{new}. Usefull for quick
renaming/translation of survey column names and by using internal object
\code{.trans_df}
}
\keyword{internal}
