% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{sanitize_colnames}
\alias{sanitize_colnames}
\title{Sanitize column names}
\usage{
sanitize_colnames(df)
}
\arguments{
\item{df}{A data frame with party names with special characters that need
to be sanitized.}
}
\value{
The input data frame with column names converted to lowercase ASCII
and party name abbreviations standardized.
}
\description{
Sanitize column names
}
\keyword{internal}
