% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrappers.R
\name{sanitize_percent}
\alias{sanitize_percent}
\title{Sanitize character vectors containing percentages}
\usage{
sanitize_percent(vec)
}
\arguments{
\item{vec}{Character vector containg percentages that can not be transformed 
to numerics b/c of special characters}
}
\description{
Sanitize character vectors containing percentages
}

