% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{sanitize_strings}
\alias{sanitize_strings}
\title{Sanitize character vector}
\usage{
sanitize_strings(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector with German umlauts replaced by their ASCII
equivalents (e.g. "oe" for "ö").
}
\description{
Substitute all German "Umlaute"
}
\keyword{internal}
