% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrappers.R
\name{scrape_wahlrecht}
\alias{scrape_wahlrecht}
\alias{scrape_ltw}
\title{Scrape surveys from wahlrecht.de}
\usage{
scrape_wahlrecht(address = "http://www.wahlrecht.de/umfragen/emnid.htm",
  parties = c("CDU", "SPD", "GRUENE", "FDP", "LINKE", "PIRATEN", "FW", "AFD",
  "SONSTIGE"))


  scrape_ltw(address = "http://www.wahlrecht.de/umfragen/landtage/niedersachsen.htm",
  parties = c("CDU", "SPD", "GRUENE", "FDP", "LINKE", "PIRATEN", "FW", "AFD",
  "SONSTIGE"))
}
\arguments{
\item{address}{http-address from which tables should be scraped.}

\item{parties}{A character vector containing names of parties to collapse.}
}
\description{
Scrapes survey tables and perfroms sanitization to output tidy data
}
