% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrappers.R
\name{scrape_austria}
\alias{scrape_austria}
\title{Import austrian survey results}
\usage{

  scrape_austria(address = "https://neuwal.com/wahlumfragen/openwal/neuwal-openwal.json")
}
\arguments{
\item{address}{URL of the JSON file}
}
\description{
Reads JSON file from neuwal.com
}
