% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrappers.R
\name{scrape_wahlrecht_api}
\alias{scrape_wahlrecht_api}
\title{Imports most recent election surveys via wahlrecht.de api}
\usage{

  scrape_wahlrecht_api(adress = "http://www.wahlrecht.de/umfragen/bundesweite.xml")
}
\arguments{
\item{adress}{http-Adress from which tables should be scraped.}
}
\description{
Imports most recent election surveys via wahlrecht.de api
}
