% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saint-lague-scheppers.R
\name{sls}
\alias{sls}
\title{Seat Distribution by Sainte-Lague/Schepers}
\usage{
sls(votes, parties, n_seats = 598L)
}
\arguments{
\item{votes}{A numeric vector giving the redistributes votes}

\item{parties}{A character vector indicating the names of parties with
respective \code{votes}.}

\item{n_seats}{The total number of seats that can be assigned to the different
parties.}
}
\value{
A named integer vector of seat counts, one entry per party,
in the same order as \code{parties}. The vector has a logical attribute
\code{ties}: \code{TRUE} if two or more parties had equal claim to the last
seat (i.e. the result is not uniquely determined and was resolved randomly),
\code{FALSE} otherwise. When \code{ties = TRUE}, re-running with a different
random seed may produce a different but equally valid seat distribution.
}
\description{
Calculates number of seats for the respective parties that have received more
than 5\% of votes (according to the method of Sainte-Lague/Schepers,
see https://www.wahlrecht.de/verfahren/rangmasszahlen.html).
}
\examples{
library(coalitions)
library(dplyr)
# get the latest survey for a sample of German federal election polls
surveys <- get_latest(surveys_sample) \%>\% ungroup() \%>\% slice(1) \%>\% tidyr::unnest("survey")
# calculate the seat distribution based on Sainte-Lague/Schepers for a parliament with 300 seats
sls(surveys$votes, surveys$party, n_seats = 300)
}
\seealso{
\code{\link{dHondt}}
}
