% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saint-lague-scheppers.R
\name{sls2}
\alias{sls2}
\title{Seat Distribution by Sainte-Lague/Scheppers}
\usage{
sls2(votes, parties, seats = 598)
}
\arguments{
\item{seats}{Number of seats in parliament. Defaults to 598 (seats in german 
parliament).}
}
\value{
A \code{data.frame} containing parties above the hurdle and the respective 
seats/percentages after redistribution via Sainte-Lague/Scheppers.
}
\description{
Calculates number of seats for the respective parties that have received more 
than 5\% of votes (according to the method of Sainte-Lague/Schepers,
see http://www.wahlrecht.de/verfahren/rangmasszahlen.html).
}
\seealso{
\code{\link{dHondt}}
}
