% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surveys_sample}
\alias{surveys_sample}
\title{Sample of selected surveys}
\format{A nested data frame with 7 rows and 2 columns:
\describe{
  \item{institute}{name of the pollster}
  \item{surveys}{a list of data frames, each containing one survey}
}}
\source{
\url{http://www.wahlrecht.de/}
}
\usage{
surveys_sample
}
\description{
A data set with surveys from seven different pollsters, three surveys per
pollster. Surveys report support for different parties in the running
for the German Bundestag prior to the 2017 election.
}
\keyword{datasets}
