% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{try_htmlSource}
\alias{try_htmlSource}
\title{Try call of read_html that throws an error if the url cannot be resolved}
\usage{
try_htmlSource(url)
}
\description{
Try call of read_html that throws an error if the url cannot be resolved
}
