% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{try_readHTML}
\alias{try_readHTML}
\title{Try call of read_html that throws an error if the url cannot be resolved}
\usage{
try_readHTML(url)
}
\arguments{
\item{url}{http-address that should be scraped.}
}
\value{
An \code{xml_document} object as returned by \code{xml2::read_html}.
}
\description{
Try call of read_html that throws an error if the url cannot be resolved
}
