% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHondt.R
\name{dHondt}
\alias{dHondt}
\title{Seat Distribution by D'Hondt}
\usage{
dHondt(votes, parties, n_seats = 183)
}
\arguments{
\item{votes}{Number of votes per party.}

\item{parties}{Names of parties (must be same length as votes).}

\item{n_seats}{Number of seats in parliament. Defaults to 183 (seats in
Austrian parliament).}
}
\value{
A \code{data.frame} containing parties above the hurdle and the respective
seats/percentages after redistribution via D'Hondt
}
\description{
Calculates number of seats for the respective parties that have received more
than \code{hurdle} percent of votes (according to the method of D'Hondt)
}
\examples{
library(coalitions)
library(dplyr) 
# scrape the newest survey from the Emnid polling agency
surveys <- get_surveys() \%>\% filter(pollster == "emnid") \%>\% tidyr::unnest() \%>\% 
  slice(1) \%>\% tidyr::unnest()
# calculate the seat distribution based on D'Hondt for a parliament with 300 seats
dHondt(surveys$votes, surveys$party, n_seats = 300)
}
\seealso{
\code{\link{sls}}
}
