% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHondt.R
\name{dHondt}
\alias{dHondt}
\title{Seat Distribution by D'Hondt}
\usage{
dHondt(survey, hurdle = 0.04, seats = 183, epsilon = 1e-06)
}
\arguments{
\item{survey}{Results of a survey as data.frame containing party names and votes.}

\item{hurdle}{The percentage of votes that must be reached to get seats in
parliament. Defaults to 0.04 (hurdle for austrian parliament).}

\item{seats}{Number of seats in parliament. Defaults to 183 (seats in austrian
parliament).}

\item{epsilon}{The percentages of votes in survey must add up to 1,
this allows for some numerical imprecission. Defaults to 10e-6.}
}
\value{
A \code{data.frame} containing parties above the hurdle and the respective
seats/percentages after redistribution via D'Hondt
}
\description{
Calculates number of seats for the respective parties that have received more
than \code{hurdle} percent of votes (according to the method of D'Hondt)
}
\seealso{
\code{\link{sls}}
}

