% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrappers.R
\name{extract_num}
\alias{extract_num}
\title{Extract numericals from string or character}
\usage{
extract_num(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Removes all characters that are not in [0-9].
}
\keyword{internal}
