% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalition-probability.R
\name{filter_superior}
\alias{filter_superior}
\title{Remove rows from table for which superior coalitions possible}
\usage{
filter_superior(majority_df, coalition, ...)
}
\arguments{
\item{majority_df}{A data frame containing logical values indicating
if the coalitions (columns) have a majority (rows).}

\item{coalition}{The coaliton of interest for which superior coalitions
will be obtained by \code{\link[coalitions]{get_superior}}.}

\item{...}{Further arguments passed to \code{\link[coalitions]{get_superior}}}
}
\description{
Remove rows from table for which superior coalitions possible
}
\seealso{
\code{\link[coalitions]{get_superior}}
}
\keyword{internal}
