% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntryProbabilities.R
\name{get_entryprobability}
\alias{get_entryprobability}
\title{Get probabilities to enter the parliament.}
\usage{
get_entryprobability(dirichlet.draws, hurdle = 0.05)
}
\arguments{
\item{dirichlet.draws}{Matrix or data frame containing draws from 
the posterior (see \code{\link{draw_from_posterior}}).}

\item{hurdle}{the percentage of votes that must be reached to get seats in 
parliament. Defaults to 0.05 (hurdle for german parliament).}
}
\value{
Vector of (named) entry probabilities.
}
\description{
Get probabilities to enter the parliament.
}
\seealso{
\code{\link{draw_from_posterior}}
}

