% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrappers.R
\name{get_meta}
\alias{get_meta}
\title{Extract "meta" information from survey data base}
\usage{
get_meta(surveys_df)
}
\arguments{
\item{surveys_df}{A data frame containing surveys from different survey
institutes as returned by \code{\link{get_surveys}}.}
}
\description{
Extract "meta" information from survey data base
}
\keyword{internal}
