% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seat-distribution.R
\name{get_seats}
\alias{get_seats}
\title{Calculate seat distribution from draws from posterior}
\usage{
get_seats(dirichlet.draws, survey, distrib.fun = sls, samplesize = NULL,
  hurdle = 0.05, others = "others", ...)
}
\arguments{
\item{dirichlet.draws}{Matrix containing random draws from posterior.}

\item{survey}{The actual survey results on which \code{dirichlet.draws}
were based on.}

\item{distrib.fun}{Function to calculate seat distribution. Defaults to
\code{\link{sls}} (Sainte-Lague/Scheppers).}

\item{samplesize}{Number of individuals participating in the \code{survey}.}

\item{hurdle}{The percentage threshold which has to be reached by a party
to enter the parliament.}

\item{others}{A string indecating the name under which parties not listed
explicitly are subsummed.}

\item{...}{Further arguments passed to \code{distrib.fun}.}
}
\value{
A data frame containing seat distributions for each simulation in
\code{dirichlet.draws}
}
\description{
Calculate seat distribution from draws from posterior
}
\seealso{
\code{\link{draw_from_posterior}}, \code{\link{sls}},
\code{\link{dHondt}}
}
\keyword{distribution}
\keyword{seat}
