% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalition-probability.R
\name{get_superior}
\alias{get_superior}
\title{Extract superior coalitions from coalition string or vector}
\usage{
get_superior(string, pattern = "_", collapse = "_")
}
\arguments{
\item{pattern}{Pattern to look for (regular expression).}

\item{collapse}{string that will be used to concatenate multiple elements
obtained by splitting \code{string} to one string.}

\item{stirng}{A character.}
}
\description{
Extract superior coalitions from coalition string or vector
}
\seealso{
stringr str_split
}
\keyword{internal}
