% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrappers.R, R/survey-helpers.R
\name{get_surveys}
\alias{get_surveys}
\alias{get_latest}
\title{Scrape surveys from all pollsters}
\usage{
get_surveys(country = c("DE", "AT"))

get_latest(surveys = NULL, max_date = Sys.Date())
}
\arguments{
\item{country}{Choose country from which surveys should be scrapped.
Currently \code{"DE"} (Germany) and \code{"AT"} (Austria) are supported.}

\item{surveys}{If provided, latest survey will be obtained from this object,
otherwise calls \code{\link{get_surveys}}.}

\item{max_date}{Specifies the date, relative to which latest survey will
be searched for. Defaults to \code{Sys.Date}.}
}
\description{
Scrape surveys from all pollsters

Given a specific date, extract the survey from this date or the last one
before this date.
}
