% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R, R/survey-helpers.R
\name{get_surveys}
\alias{get_surveys}
\alias{get_surveys_by}
\alias{get_surveys_nds}
\alias{get_surveys_saxony}
\alias{get_surveys_brb}
\alias{get_surveys_thuringen}
\alias{get_latest}
\title{Scrape surveys from all pollsters}
\usage{
get_surveys(country = c("DE", "AT"))

get_surveys_by()

get_surveys_nds()

get_surveys_saxony()

get_surveys_brb()

get_surveys_thuringen()

get_latest(surveys = NULL, max_date = Sys.Date())
}
\arguments{
\item{country}{Choose country from which surveys should be scraped.
Currently \code{"DE"} (Germany) and \code{"AT"} (Austria) are supported.}

\item{surveys}{If provided, latest survey will be obtained from this object,
otherwise calls \code{\link{get_surveys}}.}

\item{max_date}{Specifies the date, relative to which latest survey will
be searched for. Defaults to \code{Sys.Date}.}
}
\value{
Nested tibble. When fully unnested, the dataset contains the following
columns:
\describe{
  \item{pollster}{Character name of the polling institute.}
  \item{date}{Publication date of the poll.}
  \item{start, end}{Start and end date of the field period, i.e. the dates
  during which the poll was conducted.}
  \item{respondents}{Number of respondents in the poll.}
  \item{party}{Character name of an individual party.}
  \item{percent}{Percentage of respondents that chose the party. Given in
  percentage points, i.e. \code{38\%} is given as \code{38}.}
  \item{votes}{Number of respondents that chose the party.}
}
}
\description{
Given a specific date, extract the survey from this date or the last one
before this date.
}
\examples{
\dontrun{
library(coalitions)
# scrape data for the German federal election
# get_surveys()
}
library(coalitions)
### Scrape the newest poll for the German federal election
# Possibility 1: Calling get_latest without arguments scrapes surveys from the web
# Possibility 2: Use get_latest() on an already scraped dataset
surveys <- get_latest(surveys_sample)
}
