% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoalitionProbability_v3.R
\name{has_majority}
\alias{has_majority}
\title{Does a coalition have a majority}
\usage{
has_majority(seats_tab, coalition, majority = 300L)
}
\arguments{
\item{seats_tab}{A}
}
\description{
Does a coalition have a majority
}
