% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoalitionProbability_v3.R
\name{have_majority}
\alias{have_majority}
\title{Do coalitions have a majority}
\usage{
have_majority(seats_tab, coalitions = list(c("cdu"), c("cdu", "fdp"), c("cdu",
  "fdp", "gruene"), c("spd"), c("spd", "linke"), c("spd", "linke", "gruene")),
  majority = 300L, collapse = "_")
}
\arguments{
\item{seats_tab}{A data frame cotaining number of seats obtained by a party. 
Must have columns \code{party} and \code{seats}.}

\item{coalitions}{A list of coaltions for which coalition probabilities should 
be calculated. Each list entry must be a vector of party names. Those names 
need to correspond to the names in \code{majority_df}.}
}
\description{
Do coalitions have a majority
}
