% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling.R
\name{pool_austria}
\alias{pool_austria}
\title{Pool surveys from different pollsters}
\usage{
pool_austria(..., pollsters = c("Market", "Research Affairs",
  "Unique Research", "OGM", "IMAS", "Hajek", "Gallup", "Karmasin"))
}
\arguments{
\item{pollsters}{Character vector of pollsters that should be considered
for pooling.}
}
\description{
Pool surveys from different pollsters
}
\seealso{
pool_surveys
}
\keyword{internal}
