% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seat-distribution.R
\name{redistribute}
\alias{redistribute}
\title{Calculate percentage of votes/seats after excluding parties with
\code{votes < hurdle}}
\usage{
redistribute(survey, hurdle = 0.05, others = "others", epsilon = 1e-05)
}
\arguments{
\item{survey}{The actual survey results on which \code{dirichlet.draws}
were based on.}

\item{hurdle}{The percentage threshold which has to be reached by a party
to enter the parliament.}

\item{others}{A string indecating the name under which parties not listed
explicitly are subsummed.}

\item{epsilon}{Percentages should add up to 1. If they do not, within accuracy
of \code{epsilon}, an error is thrown.}
}
\description{
Calculate percentage of votes/seats after excluding parties with
\code{votes < hurdle}
}
\seealso{
\code{\link{get_seats}}, \code{\link{sls}}
}
