% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{sanitize_strings}
\alias{sanitize_strings}
\title{Sanitze character vector}
\usage{
sanitize_strings(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Substitute all german "Umlaute"
}
\keyword{internal}
